producción mp, orders 2 and recibir OC post to get

+++ db/migrate/20160501230408_create_tiempos.rb
@@ -8,7 +8,7 @@ class CreateTiempos < ActiveRecord::Migration
       t.string :Unidades
       t.integer :Costo_produccion_unitario
       t.integer :Lote_Produccion
-      t.integer :Tiempo_Medio_Producción
+      t.float :Tiempo_Medio_Producción
 
       t.timestamps null: false
     end
